Change `log.diffMerges` to `combined`
This seems like it might be more useful. The justification is that in general, what I want to know from a merge commit was whether conflicts happened, and if they did, how they were resolved. The incoming changes probably aren't that important, since I can see those from the ancestors of the merge commit.