Only provide links for GitTerm with commit context
Currently, without any commit context, file links as provided by GitTerm are not super useful and add an extra click/keypress to open the file. I figured that I'd prefer file links to only be provided with commit context (where there are more useful actions), so this switches to that behavior.