Store Fisher config files in a separate folder
This keeps the default fish config folders nice and clean.

Also, revert some changes to `init.sh` where I was relying on pushed config to install asdf programs and fisher plugins. This had the limitation of not allowing either of those to be used in `sync.sh`, which is not something I want and was already violating.