Shift in `less` by a static amount
Why would I vary horizontal shift based on width when I don't vary vertical shift based on height? A value of `20` strikes a nice balance between my eyes being able to track the content as I'm scrolling and being able to get to the end of some long content, like a table with a lot of columns.