Make features of `git history` interactively toggleable
This allows the `--graph` and `--source` flags of `git history`  to be toggled interactively with keybindings. Universal variables are used to store toggle states, which unfortunately means that cruft variables are left behind when in `git history` and killing the terminal. I'd like to figure out how to clean these up.